addressing PR review - moved helpers to helpers file, removed tiny helper
